Core/Build: Enable Clang PCH support and OS X specific features Core: Fix warnings here and there

+++ src/server/authserver/Server/RealmSocket.cpp
@@ -80,7 +80,7 @@ int RealmSocket::open(void * arg)
     return 0;
 }
 
-int RealmSocket::close(int)
+int RealmSocket::close(u_long)
 {
     shutdown();
